Setter allData-api til internal. Litt refactor.
