+++ server/src/api/endpoints/allData/allDataHandler.ts
